feat(quota): a per-principal write budget and a per-workspace storage ceiling (HD-191)
Three limiters existed, each added because someone noticed a specific expensive
path. Writes and uploads were not among them: issue, comment and attachment
creation had no per-principal budget, and while a single upload was capped at
25 MB nothing capped the total -- not per issue, not per project, not per
workspace. On Cloud the backend is S3, where we pay per byte stored and per
request made. Uploading 25 MB to S3 is more work than a search, and a search
was throttled.

The byte total lives in its own table maintained by a database trigger, not in
a SUM over attachments and not in a column on workspaces. Every non-key column
is insertable=false, updatable=false and its repository does not extend
JpaRepository, so JPA cannot write the counter at all -- the structural answer
to the issue_seq clobber, where a stale managed copy silently wrote back an
old value. The lock row is deliberately not a workspaces row.

The refusal is 409, not 429 and not 413. 413 belongs to the single-file
ceiling, 507 is a 5xx and gets auto-retried, and 429 would promise that waiting
helps. It never does: a quota does not refill, so the refusal carries no
Retry-After and a client that retries it the way it retries a 429 loops for
ever. The message prescribes nothing, because a contributor who cannot delete
attachments cannot free space, and a test asserts the absence of the words that
would dispatch them.

The reviews found the thing the design missed. Multipart resolution is eager,
so DispatcherServlet parses the part before handler mapping and therefore
before the interceptor: a body over the servlet cap mapped no handler, spent
no write unit and no bytes, and cost the node a full read plus a temp file --
the cheapest lane on the door and the only one with no bound at all. The bound
belongs at the edge, so the Caddyfile now carries request_body reading the same
variable as the servlet cap. Both misparse directions fail closed: DataSize
allows at most a two-letter unit, so a typo fails the app's boot, and an
unparseable value fails Caddy's start. There is no spelling that yields an
unbounded proxy in front of a bounded app. Since a deploy never syncs that file
to production, the drift check now asserts the block is there -- the omission
was silent, which is the failure class this ticket exists to close.

The reconciler was rewritten under measurement: the specified page-and-lock
shape ran over ten minutes holding an advisory lock against a hundred thousand
workspaces. It is now one unlocked grouped query naming only the workspaces
that disagree, then lock-and-recount for those, bounded per pass and warning
when the bound is hit. It stamps its freshness metric even when it defers to
another replica, or the staleness alert fires for ever on every multi-replica
install -- and an alert that always fires gets silenced, which would have
removed the only detector of a genuinely dead reconciler.

V26 backfills, seeds and installs the trigger in that order, because a trigger
present during the backfill makes the seed abort on its primary key. Its
one-shot half is exercised by a test that inserts rows before the migration
exists, so deleting the backfill fails the run rather than passing against an
empty table. Increments create a counter row and decrements never do: a
decrement can run inside a cascade that is removing the counter's own row, and
an insert there would fail the foreign key and break the delete.

ADR-0026. Spec in docs/design/write-budget-and-storage-quota-proposal.md.

+++ src/main/java/com/hamstrack/report/ratelimit/ReportRateLimiter.java
@@ -57,7 +57,7 @@ public class ReportRateLimiter extends PerPrincipalMinuteBudget {
     }
 
     @Override
-    protected int limit() {
+    protected long limit() {
         return reportProperties.requestsPerMinute();
     }
